feat(utils): move general purpose functions to util/utils.hpp
