+++ src/IO/Families.hpp
@@ -6,7 +6,7 @@
 struct FamilyInfo {
   std::string name;
   std::string startingGeneTree;
-  std::string ccp;
+  std::string ccpFile;
   std::string alignmentFile;
   std::string mappingFile;
   std::string libpllModel;
@@ -17,7 +17,7 @@ struct FamilyInfo {
   void reset() {
     name = "";
     startingGeneTree = "__random__";
-    ccp = "";
+    ccpFile = "";
     alignmentFile = "";
     mappingFile = "";
     libpllModel = "GTR";
